+++ src/components/datatable/components/DataTable/index.tsx
@@ -1,0 +1,18 @@
+import { forwardRef, Ref } from "react";
+import InnerDataTableWithRefs from "./components/InnerDataTable";
+import { DataTableProvider } from "../../providers";
+import { DataTableProps } from "../../types/DataTable.types";
+import { DataTableRefs } from "../../types/DataTableRef.types";
+
+const DataTableWithRefs = <FormObjectType, RecordType extends object>(
+  props: DataTableProps<FormObjectType, RecordType>,
+  ref: Ref<DataTableRefs<FormObjectType, RecordType>>,
+) => {
+  return (
+    <DataTableProvider>{InnerDataTableWithRefs(props, ref)}</DataTableProvider>
+  );
+};
+
+const DataTable = forwardRef(DataTableWithRefs);
+
+export default DataTable;
