+++ src/components/datatable/index.ts
@@ -1,1 +1,1 @@
-export { default as DataTable } from "./DataTable";
+export { default as DataTable } from "./components/DataTable";
