+++ src/components/datatable/providers/DataTableProvider/context/index.ts
@@ -1,9 +1,10 @@
+import { Context, createContext, useContext } from "react";
 import { FormInstance } from "antd";
-import { PageDefaultValues } from "../../enums/pageInfo";
+import { PageDefaultValues } from "../../../enums/pageInfo";
 import { DataTableContextType } from "../types/DataTableContext.types";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const dataContextInitData: DataTableContextType<any, any> = {
+const dataContextInitData: DataTableContextType<any, any> = {
   dataSource: [],
   updateDataSource: () => {},
   isLoading: false,
@@ -27,4 +28,18 @@ export const dataContextInitData: DataTableContextType<any, any> = {
   updatePageNumber: () => {},
   totalItems: 0,
   updateTotalItems: () => {},
+};
+
+export const DataTableContext = createContext(dataContextInitData);
+
+export const useDataTable = <FormObjectType, RecordType extends object>() => {
+  const context = useContext(
+    DataTableContext as Context<
+      DataTableContextType<FormObjectType, RecordType>
+    >,
+  );
+  if (!context) {
+    throw new Error("useDataTable must be used within a DataTableProvider");
+  }
+  return context;
 };
