+++ src/components/datatable/providers/DataTableProvider/hooks/useSearchRequest.ts
@@ -6,7 +6,7 @@ import {
 import {
   OnSearchFunction,
   OnSearchResponseData,
-} from "../../components/SearchFields/types/SearchForm.types";
+} from "../../../components/SearchFields/types/SearchForm.types";
 
 const useSearchRequest = <
   FormObjectType,
