+++ src/components/datatable/providers/DataTableProvider/provider/index.tsx
@@ -1,28 +1,10 @@
-import { Context, createContext, useContext } from "react";
-import {
-  DataTableContextType,
-  DataTableProviderProps,
-} from "./types/DataTableContext.types";
-import useDataSource from "./hooks/useDataSource";
-import useLoading from "./hooks/useLoading";
-import useSearchRequest from "./hooks/useSearchRequest";
-import useSearchForm from "./hooks/useSearchForm";
-import usePageInfo from "./hooks/usePageInfo";
-import { dataContextInitData } from "./data/dataTableContextInitData";
-
-const DataTableContext = createContext(dataContextInitData);
-
-export const useDataTable = <FormObjectType, RecordType extends object>() => {
-  const context = useContext(
-    DataTableContext as Context<
-      DataTableContextType<FormObjectType, RecordType>
-    >,
-  );
-  if (!context) {
-    throw new Error("useDataTable must be used within a DataTableProvider");
-  }
-  return context;
-};
+import { DataTableContext } from "../context";
+import useDataSource from "../hooks/useDataSource";
+import useLoading from "../hooks/useLoading";
+import usePageInfo from "../hooks/usePageInfo";
+import useSearchForm from "../hooks/useSearchForm";
+import useSearchRequest from "../hooks/useSearchRequest";
+import { DataTableProviderProps } from "../types/DataTableContext.types";
 
 export const DataTableProvider = <FormObjectType, RecordType extends object>({
   children,
