+++ src/components/datatable/providers/index.ts
@@ -1,0 +1,1 @@
+export * from "./DataTableProvider";
