+++ src/components/datatable/types/index.ts
@@ -1,0 +1,4 @@
+export type * from "./DataTable.types";
+export type * from "./DataTableRef.types";
+export type * from "./useDataSource.types";
+export type * from "./useRequestData.types";
